requirements.txt
Flask==2.3.3
opencv-python==4.8.0.74
numpy==1.24.3
qrcode==7.4.2
pyzbar==0.1.9
Pillow==10.0.0
python-dateutil==2.8.2
opencv-contrib-python==4.8.0.74